Move configs and data files
